Refractored the entire project. Removed unneccessary files
