+++ src/components/button/KsButton.tsx
@@ -1,5 +1,4 @@
 import react from 'react';
-import "./styles.scss";
 import { Button, ButtonProps, Icon } from 'trunx';
 import { faThumbsUp } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
@@ -8,9 +7,9 @@ interface KsButtonProperties extends ButtonProps {
     buttonText: string;
 }
 
-export default function KsButton({buttonText}: KsButtonProperties){
+export default function KsButton({ buttonText }: KsButtonProperties) {
 
-    return(
+    return (
         <Button isLink>
             <Icon className="iconHolder">
                 <FontAwesomeIcon icon={faThumbsUp} />
@@ -20,5 +19,5 @@ export default function KsButton({buttonText}: KsButtonProperties){
     )
 }
 
-{/*Hvordan sende inn noe som krever en import? Her lurer jeg på hvordan jeg ville ha sendt inn noe til KsButton som endret hvilke icon som skulle være der */}
-{/* hvorfor må jeg bruke spread operatoren på type her? */}
+{/*Hvordan sende inn noe som krever en import? Her lurer jeg på hvordan jeg ville ha sendt inn noe til KsButton som endret hvilke icon som skulle være der */ }
+{/* hvorfor må jeg bruke spread operatoren på type her? */ }
