--- src/pages/RegisterUserTypePage.scss
@@ -1,1 +1,0 @@
-
